fake/fake-backend/requirements.txt
Flask
gunicorn
python-dotenv
cloudinary
sendgrid
pytz
apscheduler
moviepy
opencv-python
torch
torchvision
Pillow
